# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Express TypeScript Boilerplate API
  description: |
    Production-ready Express.js TypeScript boilerplate for blockchain microservices.

    This API provides comprehensive user management, authentication, and health monitoring
    capabilities with enterprise-grade security features.

    ## Features
    - JWT and API Key authentication
    - Role-based access control (RBAC)
    - Rate limiting with Redis backend
    - Input validation and sanitization
    - Comprehensive health checks
    - Metrics collection
    - Security headers and CORS

    ## Authentication
    This API supports two authentication methods:
    1. **JWT Bearer Token** - For user sessions
    2. **API Key** - For service-to-service communication

    ## Rate Limiting
    All endpoints are rate limited. Limits vary by endpoint and authentication method.
    Rate limit information is provided in response headers.

    ## Error Handling
    All errors follow a consistent format with appropriate HTTP status codes.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
  - url: https://api.example.com/v1
    description: Production server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key in format "keyId.key"
  parameters:
    PageParam:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
        example: 1
    LimitParam:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
        example: 10
    SortByParam:
      name: sortBy
      in: query
      description: Field to sort by
      schema:
        type: string
        example: createdAt
    SortOrderParam:
      name: sortOrder
      in: query
      description: Sort order
      schema:
        type: string
        enum:
          - asc
          - desc
        default: desc
        example: desc
  schemas:
    ApiResponse:
      type: object
      required:
        - success
        - meta
      properties:
        success:
          type: boolean
          example: true
        data:
          description: Response data
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
            message:
              type: string
              example: Validation failed
            details:
              description: Additional error details
        meta:
          type: object
          properties:
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
            version:
              type: string
              example: v1
    PaginatedResponse:
      allOf:
        - $ref: '#/components/schemas/ApiResponse'
        - type: object
          properties:
            pagination:
              type: object
              properties:
                page:
                  type: integer
                  example: 1
                limit:
                  type: integer
                  example: 10
                total:
                  type: integer
                  example: 100
                totalPages:
                  type: integer
                  example: 10
    User:
      type: object
      properties:
        id:
          type: string
          example: clp1234567890abcdef
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: john_doe
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreateUserRequest:
      type: object
      required:
        - email
        - username
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        username:
          type: string
          example: john_doe
        password:
          type: string
          minLength: 8
          example: SecurePassword123!
        firstName:
          type: string
          example: John
        lastName:
          type: string
          example: Doe
    UpdateUserRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        username:
          type: string
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        isActive:
          type: boolean
...
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: VALIDATION_ERROR
                      message:
                        type: string
                        example: Validation failed
                      details:
                        type: array
                        items:
                          type: object
                          properties:
                            field:
                              type: string
                              example: email
                            message:
                              type: string
                              example: Invalid email format
    UnauthorizedError:
      description: Authentication required
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: UNAUTHORIZED
                      message:
                        type: string
                        example: Authentication required
    ForbiddenError:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: FORBIDDEN
                      message:
                        type: string
                        example: Insufficient permissions
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: NOT_FOUND
                      message:
                        type: string
                        example: Resource not found
    ConflictError:
      description: Resource conflict
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: CONFLICT
                      message:
                        type: string
                        example: Resource already exists
    RateLimitError:
      description: Rate limit exceeded
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/ApiResponse'
              - type: object
                properties:
                  success:
                    type: boolean
                    example: false
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                        example: RATE_LIMIT_EXCEEDED
                      message:
                        type: string
                        example: Rate limit exceeded
paths:
...
  /users:
    get:
      tags:
        - Users
      summary: List users with pagination and filtering
      description: |
        Retrieve a paginated list of users with optional filtering.
        Requires authentication and appropriate permissions.
      operationId: getUsers
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - $ref: ../base.yaml#/components/parameters/PageParam
        - $ref: ../base.yaml#/components/parameters/LimitParam
        - $ref: ../base.yaml#/components/parameters/SortByParam
        - $ref: ../base.yaml#/components/parameters/SortOrderParam
        - name: search
          in: query
          description: Search term to filter users by name or email
          schema:
            type: string
            minLength: 1
            maxLength: 100
            example: john
        - name: isActive
          in: query
          description: Filter users by active status
          schema:
            type: boolean
            example: true
      responses:
        '200':
          description: Users retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../base.yaml#/components/schemas/PaginatedResponse
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: ../base.yaml#/components/schemas/User
        '400':
          $ref: ../base.yaml#/components/responses/ValidationError
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
        '403':
          $ref: ../base.yaml#/components/responses/ForbiddenError
        '429':
          $ref: ../base.yaml#/components/responses/RateLimitError
    post:
      tags:
        - Users
      summary: Create a new user
      description: |
        Create a new user account. This operation requires admin privileges.
        The password will be hashed before storage.
      operationId: createUser
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../base.yaml#/components/schemas/CreateUserRequest
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../base.yaml#/components/schemas/ApiResponse
                  - type: object
                    properties:
                      data:
                        $ref: ../base.yaml#/components/schemas/User
        '400':
          $ref: ../base.yaml#/components/responses/ValidationError
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
        '403':
          $ref: ../base.yaml#/components/responses/ForbiddenError
        '409':
          $ref: ../base.yaml#/components/responses/ConflictError
        '429':
          $ref: ../base.yaml#/components/responses/RateLimitError
  /users/me:
    get:
      tags:
        - Users
      summary: Get current user profile
      description: |
        Retrieve the profile of the currently authenticated user.
        This endpoint allows users to view their own profile information.
      operationId: getCurrentUser
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      responses:
        '200':
          description: Current user profile retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../base.yaml#/components/schemas/ApiResponse
                  - type: object
                    properties:
                      data:
                        $ref: ../base.yaml#/components/schemas/User
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
    put:
      tags:
        - Users
      summary: Update current user profile
      description: |
        Update the profile of the currently authenticated user.
        Users can only update their own profile information.
      operationId: updateCurrentUser
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../base.yaml#/components/schemas/UpdateUserRequest
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../base.yaml#/components/schemas/ApiResponse
                  - type: object
                    properties:
                      data:
                        $ref: ../base.yaml#/components/schemas/User
        '400':
          $ref: ../base.yaml#/components/responses/ValidationError
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
        '409':
          $ref: ../base.yaml#/components/responses/ConflictError
        '429':
          $ref: ../base.yaml#/components/responses/RateLimitError
  /users/{id}:
    get:
      tags:
        - Users
      summary: Get user by ID
      description: |
        Retrieve a specific user by their ID.
        Users can view their own profile, admins can view any user.
      operationId: getUserById
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
            example: clp1234567890abcdef
      responses:
        '200':
          description: User retrieved successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../base.yaml#/components/schemas/ApiResponse
                  - type: object
                    properties:
                      data:
                        $ref: ../base.yaml#/components/schemas/User
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
        '403':
          $ref: ../base.yaml#/components/responses/ForbiddenError
        '404':
          $ref: ../base.yaml#/components/responses/NotFoundError
    put:
      tags:
        - Users
      summary: Update user by ID
      description: |
        Update a specific user by their ID.
        Users can update their own profile, admins can update any user.
      operationId: updateUserById
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
            example: clp1234567890abcdef
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: ../base.yaml#/components/schemas/UpdateUserRequest
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                allOf:
                  - $ref: ../base.yaml#/components/schemas/ApiResponse
                  - type: object
                    properties:
                      data:
                        $ref: ../base.yaml#/components/schemas/User
        '400':
          $ref: ../base.yaml#/components/responses/ValidationError
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
        '403':
          $ref: ../base.yaml#/components/responses/ForbiddenError
        '404':
          $ref: ../base.yaml#/components/responses/NotFoundError
        '409':
          $ref: ../base.yaml#/components/responses/ConflictError
        '429':
          $ref: ../base.yaml#/components/responses/RateLimitError
    delete:
      tags:
        - Users
      summary: Delete user by ID
      description: |
        Delete a specific user by their ID.
        This operation requires admin privileges and cannot be undone.
      operationId: deleteUserById
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: User ID
          schema:
            type: string
            format: uuid
            example: clp1234567890abcdef
      responses:
        '204':
          description: User deleted successfully
        '401':
          $ref: ../base.yaml#/components/responses/UnauthorizedError
        '403':
          $ref: ../base.yaml#/components/responses/ForbiddenError
        '404':
          $ref: ../base.yaml#/components/responses/NotFoundError
        '429':
          $ref: ../base.yaml#/components/responses/RateLimitError